Fix modal instrument note state and tremolo rate

+++ Opcodes/marimba.h
@@ -47,8 +47,7 @@ typedef struct Marimba {
     MYFLT       strikePosition;
     MYFLT       stickHardness;
     int32_t         first;
-    int32_t
-    kloop;
+    double      kloop;
 } MARIMBA;
 
 #endif
